+++ src/tikray/core.py
@@ -1,8 +1,9 @@
-import json
 import logging
 import sys
 import typing as t
 from pathlib import Path
+
+import orjson as json
 
 from tikray.model.collection import CollectionAddress, CollectionTransformation
 from tikray.model.project import TransformationProject
@@ -25,8 +26,8 @@ def process_project(transformation: Path, input_: Path, output: Path):
             continue
         data = json.loads(Path(item).read_text())
         output_path = output / item.name
-        with open(output_path, "w") as output_stream:
-            print(jd(tikray_transformation.apply(data)), file=output_stream)
+        with open(output_path, "wb") as output_stream:
+            output_stream.write(jd(tikray_transformation.apply(data)))
             logger.info(f"Processed output: {output_path}")
 
 
@@ -35,10 +36,10 @@ def process_collection(transformation: Path, input_: Path, output: t.Optional[Pa
     data = json.loads(input_.read_text())
     ct = CollectionTransformation.from_yaml(transformation.read_text())
     result = ct.apply(data)
-    output_stream = sys.stdout
+    output_stream = sys.stdout.buffer
     if output is not None:
         if output.is_dir():
             output = output / input_.name
-        output_stream = open(output, "w")
+        output_stream = open(output, "wb")
     output_stream.write(jd(result))
     output_stream.flush()